add prepare_generate_proof
