+++ zokrates_core/src/proof_system/utils.rs
@@ -1,9 +1,10 @@
 use field::Field;
 use flat_absy::flat_variable::FlatVariable;
+use std::cmp::max;
 use std::ffi::CString;
 
 // utility function. Converts a Fields vector-based byte representation to fixed size array.
-pub fn vec_as_u8_32_array(vec: &Vec<u8>) -> [u8; 32] {
+fn vec_as_u8_32_array(vec: &Vec<u8>) -> [u8; 32] {
     assert!(vec.len() <= 32);
     let mut array = [0u8; 32];
     for (index, byte) in vec.iter().enumerate() {
@@ -141,5 +142,40 @@ pub fn prepare_setup<T: Field>(
         num_inputs,
         pk_path_cstring,
         vk_path_cstring,
+    )
+}
+
+// proof-system-independent preparation for proof generation
+pub fn prepare_generate_proof<T: Field>(
+    pk_path: &str,
+    proof_path: &str,
+    public_inputs: Vec<T>,
+    private_inputs: Vec<T>,
+) -> (CString, CString, Vec<[u8; 32]>, usize, Vec<[u8; 32]>, usize) {
+    let pk_path_cstring = CString::new(pk_path).unwrap();
+    let proof_path_cstring = CString::new(proof_path).unwrap();
+
+    let public_inputs_length = public_inputs.len();
+    let private_inputs_length = private_inputs.len();
+
+    let mut public_inputs_arr: Vec<[u8; 32]> = vec![[0u8; 32]; public_inputs_length];
+    // length must not be zero here, so we apply the max function
+    let mut private_inputs_arr: Vec<[u8; 32]> = vec![[0u8; 32]; max(private_inputs_length, 1)];
+
+    //convert inputs
+    for (index, value) in public_inputs.into_iter().enumerate() {
+        public_inputs_arr[index] = vec_as_u8_32_array(&value.into_byte_vector());
+    }
+    for (index, value) in private_inputs.into_iter().enumerate() {
+        private_inputs_arr[index] = vec_as_u8_32_array(&value.into_byte_vector());
+    }
+
+    (
+        pk_path_cstring,
+        proof_path_cstring,
+        public_inputs_arr,
+        public_inputs_length,
+        private_inputs_arr,
+        private_inputs_length,
     )
 }
